refactor: HttpHeaders를 클래스로 변경
