+++ tomcat/src/main/java/org/apache/http/Cookie.java
@@ -1,4 +1,4 @@
-package org.apache.cookie;
+package org.apache.http;
 
 import java.util.Arrays;
 import java.util.HashMap;
