+++ tomcat/src/main/java/org/apache/http/HttpHeaders.java
@@ -1,0 +1,9 @@
+package org.apache.http;
+
+public final class HttpHeaders {
+
+    public static final String LOCATION = "Location";
+    public static final String SET_COOKIE = "Set-Cookie";
+    public static final String CONTENT_TYPE = "Content-Type";
+    public static final String CONTENT_LENGTH = "Content-Length";
+}
